requirements.txt
alembic==1.9.4
antlr4-python3-runtime==4.9.3
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
cloudpickle==2.2.1
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
databricks-cli==0.17.4
debugpy==1.6.6
decorator==5.1.1
docker==6.0.1
entrypoints==0.4
executing==1.2.0
Flask==2.2.3
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-SQLAlchemy==3.0.3
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
gunicorn==20.1.0
Hydra==2.5
hydra-core==1.3.1
idna==3.4
imbalanced-learn==0.10.1
imblearn==0.0
importlib-metadata==5.2.0
importlib-resources==5.12.0
ipykernel==6.21.2
ipython==8.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter-client==8.0.3
jupyter-core==5.2.0
kiwisolver==1.4.4
llvmlite==0.39.1
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
marshmallow==3.19.0
matplotlib==3.7.0
matplotlib-inline==0.1.6
mlflow==2.1.1
mysqlclient==2.1.1
nest-asyncio==1.5.6
numba==0.56.4
numpy==1.24.2
oauthlib==3.2.2
omegaconf==2.3.0
packaging==22.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
prompt-toolkit==3.0.38
protobuf==4.22.0
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
Pygments==2.14.0
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
querystring-parser==1.2.4
requests==2.28.2
scipy==1.10.1
shap==0.41.0
six==1.16.0
sklearn==0.0.post1
slicer==0.0.7
smmap==5.0.0
SQLAlchemy==1.4.46
sqlparse==0.4.3
stack-data==0.6.2
tabulate==0.9.0
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
urllib3==1.26.14
wcwidth==0.2.6
websocket-client==1.5.1
Werkzeug==2.2.3
zipp==3.14.0
